refactor: DTO파일 다시 쪼개서 올렷다 민지야
