+++ Jumpit/Network/Request/ResumeDataPostRequestDTO.swift
@@ -1,0 +1,13 @@
+//
+//  ResumeDataPostRequestDTO.swift
+//  Jumpit
+//
+//  Created by 이지훈 on 5/24/24.
+//
+
+import Foundation
+
+struct ResumeDataPostRequest: Codable {
+    let title: String
+    let userId: Int
+}
