+++ Jumpit/Network/Response/PrivateResumeResponseDTO.swift
@@ -12,10 +12,3 @@ struct PrivateResumeResponse: Codable {
     let status: Int
     let message: String
 }
-
-
-struct ResumeDataPostResponse: Codable {
-    let status: Int
-    let message: String
-    let data: Data?
-}
